% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayCreation.R
\name{mZeros}
\alias{mZeros}
\alias{ArrayCreation}
\alias{mOnes}
\alias{mTRUE}
\alias{mFALSE}
\alias{mNULL}
\title{Array Creation}
\usage{
mZeros(...)

mOnes(...)

mTRUE(...)

mFALSE(...)

mNULL(...)
}
\arguments{
\item{...}{Numbers or numeric vectors, passed to \code{\link[=c]{c()}} and then used as \code{dim}
argument in a call to \code{\link[=array]{array()}}.}
}
\value{
The result of \code{array(XXX, c(...))}, where XXX is \code{0}, \code{1}, \code{\link{TRUE}}, \code{\link{FALSE}}, or \code{\link[=list]{list()}}, respectively.
}
\description{
Convenience functions that create an array full of identical entries.
}
\examples{
mZeros(2, 3)

mOnes(c(1, 2, 3))

mTRUE(c(1, 3), 2)

mFALSE(5)

}
