% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implicitExpansion.R
\name{expandArray}
\alias{expandArray}
\title{Expand an Array to a Given Dimension}
\usage{
expandArray(x, dy)
}
\arguments{
\item{x}{An object that is coerced to array.}

\item{dy}{The dimensions it is to be expanded to.}
}
\value{
An array that consists of the entries in \code{x}, with dimension \code{dy}.
}
\description{
Expand an array to a given dimension by repeating its entries in the directions
where the array has dimension one.
}
\details{
Throws an error if the array and dimensions are not compatible.
}
\examples{
x <- 1:3
expandArray(x, c(3,4))

}
\seealso{
\code{\link{expandedDim}}, \code{\link{mmapply}}
}
